[Redux] Move payee states from queriesSlice to payeesSlice
