+++ packages/desktop-client/src/components/rules/RuleEditor.tsx
@@ -75,7 +75,7 @@ import {
   SelectedProvider,
 } from '@desktop-client/hooks/useSelected';
 import { aqlQuery } from '@desktop-client/queries/aqlQuery';
-import { getPayees } from '@desktop-client/queries/queriesSlice';
+import { getPayees } from '@desktop-client/payees/payeesSlice';
 import { useDispatch } from '@desktop-client/redux';
 import { enableUndo, disableUndo } from '@desktop-client/undo';
