+++ packages/desktop-client/src/components/transactions/TransactionsTable.tsx
@@ -126,10 +126,9 @@ import {
 import { pushModal } from '@desktop-client/modals/modalsSlice';
 import { NotesTagFormatter } from '@desktop-client/notes/NotesTagFormatter';
 import { addNotification } from '@desktop-client/notifications/notificationsSlice';
-import {
-  getPayeesById,
-  getCategoriesById,
-} from '@desktop-client/queries/queriesSlice';
+import { getAccountsById } from '@desktop-client/accounts/accountsSlice';
+import { getPayeesById } from '@desktop-client/payees/payeesSlice';
+import { getCategoriesById } from '@desktop-client/queries/queriesSlice';
 import { useDispatch } from '@desktop-client/redux';
 
 type TransactionHeaderProps = {
