+++ packages/desktop-client/src/hooks/usePayees.ts
@@ -5,14 +5,14 @@ import { useInitialMount } from './useInitialMount';
 import {
   getCommonPayees,
   getPayees,
-} from '@desktop-client/queries/queriesSlice';
+} from '@desktop-client/payees/payeesSlice';
 import { useSelector, useDispatch } from '@desktop-client/redux';
 
 export function useCommonPayees() {
   const dispatch = useDispatch();
   const isInitialMount = useInitialMount();
   const isCommonPayeesDirty = useSelector(
-    state => state.queries.isCommonPayeesDirty,
+    state => state.payees.isCommonPayeesDirty,
   );
 
   useEffect(() => {
@@ -21,13 +21,13 @@ export function useCommonPayees() {
     }
   }, [dispatch, isInitialMount, isCommonPayeesDirty]);
 
-  return useSelector(state => state.queries.commonPayees);
+  return useSelector(state => state.payees.commonPayees);
 }
 
 export function usePayees() {
   const dispatch = useDispatch();
   const isInitialMount = useInitialMount();
-  const isPayeesDirty = useSelector(state => state.queries.isPayeesDirty);
+  const isPayeesDirty = useSelector(state => state.payees.isPayeesDirty);
 
   useEffect(() => {
     if (isInitialMount || isPayeesDirty) {
@@ -35,5 +35,5 @@ export function usePayees() {
     }
   }, [dispatch, isInitialMount, isPayeesDirty]);
 
-  return useSelector(state => state.queries.payees);
+  return useSelector(state => state.payees.payees);
 }
