+++ packages/desktop-client/src/redux/mock.tsx
@@ -26,6 +26,10 @@ import {
   reducer as notificationsSliceReducer,
 } from '@desktop-client/notifications/notificationsSlice';
 import {
+  name as payeesSliceName,
+  reducer as payeesSliceReducer,
+} from '@desktop-client/payees/payeesSlice';
+import {
   name as prefsSliceName,
   reducer as prefsSliceReducer,
 } from '@desktop-client/prefs/prefsSlice';
@@ -44,6 +48,7 @@ const appReducer = combineReducers({
   [budgetsSliceName]: budgetsSliceReducer,
   [modalsSliceName]: modalsSliceReducer,
   [notificationsSliceName]: notificationsSliceReducer,
+  [payeesSliceName]: payeesSliceReducer,
   [prefsSliceName]: prefsSliceReducer,
   [queriesSliceName]: queriesSliceReducer,
   [usersSliceName]: usersSliceReducer,
