+++ packages/desktop-client/src/redux/store.ts
@@ -27,6 +27,10 @@ import {
   addNotification,
 } from '@desktop-client/notifications/notificationsSlice';
 import {
+  name as payeesSliceName,
+  reducer as payeesSliceReducer,
+} from '@desktop-client/payees/payeesSlice';
+import {
   name as prefsSliceName,
   reducer as prefsSliceReducer,
 } from '@desktop-client/prefs/prefsSlice';
@@ -45,6 +49,7 @@ const rootReducer = combineReducers({
   [budgetsSliceName]: budgetsSliceReducer,
   [modalsSliceName]: modalsSliceReducer,
   [notificationsSliceName]: notificationsSliceReducer,
+  [payeesSliceName]: payeesSliceReducer,
   [prefsSliceName]: prefsSliceReducer,
   [queriesSliceName]: queriesSliceReducer,
   [usersSliceName]: usersSliceReducer,
