+++ packages/desktop-client/src/sync-events.ts
@@ -11,8 +11,9 @@ import {
   addNotification,
   type Notification,
 } from './notifications/notificationsSlice';
+import { reloadPayees } from './payees/payeesSlice';
 import { loadPrefs } from './prefs/prefsSlice';
-import { reloadCategories, reloadPayees } from './queries/queriesSlice';
+import { reloadCategories } from './queries/queriesSlice';
 import { type AppStore } from './redux/store';
 import { signOut } from './users/usersSlice';
