+++ upcoming-release-notes/5592.md
@@ -1,0 +1,6 @@
+---
+category: Maintenance
+authors: [joel-jeremy]
+---
+
+[Redux] Move payee states from queriesSlice to payeesSlice
